🏠 Tidy up internal property domain
