+++ packages/pbt/src/Property/index.ts
@@ -1,6 +1,4 @@
 export * from './Imports';
-export * from './PropertyIteration';
-export * from './PropertyFunction';
-export * from './reproduce';
+export * from './Reproduction';
+export * from './Exploration';
 export * from './Property';
-export * from './PropertyResult';
